add: append header to files that already exist

+++ create_config_file/notes.py
@@ -6,6 +6,8 @@ LastEditTime: 2022.02.06 16:43
 """
 import os
 import time
+from pathlib import Path
+from typing import Union
 
 from project_scaffold.notes import TEMPLATE_ARTICLE_CONTENT, TEMPLATE_ARTICLE_SETTINGS
 
@@ -22,14 +24,7 @@ url:  "{url}"  # 永久链接
 """
 
 
-def notes(path):
-    if os.path.exists(path):
-        return
-
-    parent = os.path.dirname(path)
-    if parent and not os.path.exists(parent):
-        os.makedirs(parent, exist_ok=True)
-
+def generate_notes_header(path, only_header=False):
     cwd = os.path.abspath(os.getcwd())
     prefix = os.path.basename(cwd)
 
@@ -54,7 +49,7 @@ def notes(path):
 
     template_article += NOTE_HEADER_SEPARATOR + "\n"
 
-    if os.path.isfile(TEMPLATE_ARTICLE_CONTENT.file):
+    if not only_header and os.path.isfile(TEMPLATE_ARTICLE_CONTENT.file):
         with open(TEMPLATE_ARTICLE_CONTENT.file, encoding="utf-8") as fp:
             content = fp.read()
             if content:
@@ -62,5 +57,40 @@ def notes(path):
             else:
                 template_article += (TEMPLATE_ARTICLE_CONTENT.content + "\n\n") * 5
 
+    return template_article
+
+
+def notes(path):
+    if os.path.exists(path):
+        return
+
+    parent = os.path.dirname(path)
+    if parent and not os.path.exists(parent):
+        os.makedirs(parent, exist_ok=True)
+
+    template_article = generate_notes_header(path)
+
     with open(path, "w", encoding="utf-8", newline="\n") as fp:
         fp.write(template_article)
+
+
+def notes_append_header(path: Union[Path, str]):
+    """添加到已存在的笔记"""
+
+    path = Path(path)
+
+    if not path.exists():
+        return
+
+    if path.is_dir():
+        for entry in path.iterdir():
+            notes_append_header(entry)
+    elif path.suffix == ".md":
+        if path.open().read(3) == "---":
+            return
+
+        template_article = generate_notes_header(path, only_header=True)
+        old_content = path.read_text(encoding="utf-8")
+
+        with open(path, "w", encoding="utf-8", newline="\n") as fp:
+            fp.write(template_article + old_content)
